last update

+++ app/Http/Controllers/CommentsController.php
@@ -91,7 +91,7 @@ class CommentsController extends Controller
 
     public function deleteComment($comment_id, $location){
         $comment = DB::select("SELECT * FROM comments_log  WHERE comment_id = '$comment_id'");
-
+        
         if(!$comment){
             $msg = 'comment does not exist!<br> Request sent by <b> @'.Auth::user()->name. '</b>';
             return view('potato')->with('msg', $msg);            
@@ -105,10 +105,11 @@ class CommentsController extends Controller
         DB::delete("DELETE FROM comments_log  WHERE comment_id = '$comment_id'");
         
         $noti_id = DB::select("SELECT * FROM comment_notifications WHERE comment_track_id = '$comment_id'");
-        $noti_id = $noti_id[0]->comment_noti_id;
-        
-        DB::delete("DELETE FROM comment_notifications WHERE comment_noti_id = '$noti_id'");
-        DB::delete("DELETE FROM notifications_log WHERE notification_id = '$noti_id'");
+        if($noti_id){
+            $noti_id = $noti_id[0]->comment_noti_id;
+            DB::delete("DELETE FROM comment_notifications WHERE comment_noti_id = '$noti_id'");
+            DB::delete("DELETE FROM notifications_log WHERE notification_id = '$noti_id'");
+        }
         
         if($location == "home"){
             return redirect()->to("/home");
